# server/swagger.yaml
openapi: 3.0.0
info:
  description: |
    Documentation
  version: '1.0'
  title: 'Coderslab'
  termsOfService: 'http://swagger.io/terms/'
  contact:
    name: Codando Community
    email: codando.community@gmail.com
    url: https://github.com/codando-community/
#  license:
#    name: Apache 2.0
#    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  /user/create:
    post:
      tags: 
        - User
      summary: Cria um novo usuário
      description: Cria um novo perfil para o usuário que precisa acessar a aplicação pela primeira vez.
      operationId: user_create
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user_read_token.Response'
        '400':
          description: Bad Request
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '406':
          description: Not Acceptable
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user.Request'
        required: true 
...
  /user/update/{id}:
    put:
      tags:
        - User
      summary: Atualiza um usuário com base em um ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID do usuário gerado pelo MongoDB
          schema:
            type: string
            format: ObjectId
            minimum: 1
      description: Atualiza os dados de perfil de um usuário e sua lista de interesses com base em seu ID gerado pelo MongoDB. Caso haja um interesse novo, ele deve vir com _id = "", caso o campo esteja preenchido, é considerado como um interesse já existente que tem que ser atualizado.
      operationId: user_update
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user_read.Response'
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '406':
          description: Not Acceptable
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user_update.Request'
        required: true
...
servers:
  - url: 'https://cluster.prod.com.br/coderslab'
    description: Link do cluster produtivo - External URL PRD
components:
  schemas:
...
    user.Request:
      title: user.Request
      description: 'Objeto para usuário'
      type: object
      properties:
        username:
          type: string
          description: 'Username do usuário Ex: Joãozinho10'
        name:
          type: string
          description: 'Nome do usuário Ex: João'
        last_name:
          type: string
          description: 'Sobrenome do usuário Ex: Silva'
        email:
          type: string
          description: 'E-mail do usuário Ex: joao@email.com'
        discord_id:
          type: string
          description: 'ID do usuário no Discord Ex: joao#0134'
        github_id:
          type: string
          description: 'ID do usuário no GitHub Ex: joaosilva10'
        password:
          type: string
          description: 'Senha escolhida pelo usuário para utilizar na plataforma Ex: Jo@10Silv4'
        interest_list:
          type: array
          items:
            type: object
            properties:
              _id_subject:
                type: string
                description: 'ID do interesse gerado pelo MongoDB'
              level:
                type: number
                description: 'Número de 0 a 2 que represente o nível de conhecimentodo grupo Ex: 1'
    user_update.Request:
      title: user.Request
      description: 'Objeto para usuário'
      type: object
      properties:
        username:
          type: string
          description: 'Username do usuário Ex: Joãozinho10'
        name:
          type: string
          description: 'Nome do usuário Ex: João'
        last_name:
          type: string
          description: 'Sobrenome do usuário Ex: Silva'
        email:
          type: string
          description: 'E-mail do usuário Ex: joao@email.com'
        discord_id:
          type: string
          description: 'ID do usuário no Discord Ex: joao#0134'
        github_id:
          type: string
          description: 'ID do usuário no GitHub Ex: joaosilva10'
        password:
          type: string
          description: 'Senha escolhida pelo usuário para utilizar na plataforma Ex: Jo@10Silv4'
        interest_list:
          type: array
          items:
            type: object
            properties:
              _id:
                type: string
                description: 'ID do interesse gerado pelo MongoDB'
              _id_user:
                type: string
                description: 'ID do usuário gerado pelo MongoDB'
              _id_subject:
                type: string
                description: 'ID do interesse gerado pelo MongoDB'
              level:
                type: number
                description: 'Número de 0 a 2 que represente o nível de conhecimentodo grupo Ex: 1'
...
    user_read.Response:
      title: user_read.Response
      description: 'Objeto para usuário'
      type: object
      properties:
        _id:
          type: string
          description: 'ID gerado pelo MongoDB'
        username:
          type: string
          description: 'Username do usuário Ex: Joãozinho10'
        name:
          type: string
          description: 'Nome do usuário Ex: João'
        last_name:
          type: string
          description: 'Sobrenome do usuário Ex: Silva'
        email:
          type: string
          description: 'E-mail do usuário Ex: joao@email.com'
        discord_id:
          type: string
          description: 'ID do usuário no Discord Ex: joao#0134'
        github_id:
          type: string
          description: 'ID do usuário no GitHub Ex: joaosilva10'
        password:
          type: string
          description: 'Senha escolhida pelo usuário para utilizar na plataforma Ex: Jo@10Silv4'
        interest_list:
          type: array
          items:
            type: object
            properties:
              _id:
                type: string
                description: 'ID do interesse gerado pelo MongoDB'
              _id_user:
                type: string
                description: 'ID do usuário gerado pelo MongoDB'
              _id_subject:
                type: string
                description: 'ID do interesse gerado pelo MongoDB'
              level:
                type: number
                description: 'Número de 0 a 2 que represente o nível de conhecimentodo grupo Ex: 1'
    user_read_token.Response:
      title: user_read.Response
      description: 'Objeto para usuário'
      type: object
      properties:
        _id:
          type: string
          description: 'ID gerado pelo MongoDB'
        username:
          type: string
          description: 'Username do usuário Ex: Joãozinho10'
        name:
          type: string
          description: 'Nome do usuário Ex: João'
        last_name:
          type: string
          description: 'Sobrenome do usuário Ex: Silva'
        email:
          type: string
          description: 'E-mail do usuário Ex: joao@email.com'
        discord_id:
          type: string
          description: 'ID do usuário no Discord Ex: joao#0134'
        github_id:
          type: string
          description: 'ID do usuário no GitHub Ex: joaosilva10'
        password:
          type: string
          description: 'Senha escolhida pelo usuário para utilizar na plataforma Ex: Jo@10Silv4'
        token:
          type: string
          description: 'Token da sessão'
        interest_list:
          type: array
          items:
            type: object
            properties:
              _id:
                type: string
                description: 'ID do interesse gerado pelo MongoDB'
              _id_user:
                type: string
                description: 'ID do usuário gerado pelo MongoDB'
              _id_subject:
                type: string
                description: 'ID do interesse gerado pelo MongoDB'
              level:
                type: number
                description: 'Número de 0 a 2 que represente o nível de conhecimentodo grupo Ex: 1'               
...